+++ src/WSI/VkSurfaceKHR.cpp
@@ -34,7 +34,7 @@ void SurfaceKHR::getSurfaceCapabilities(VkSurfaceCapabilitiesKHR *pSurfaceCapabi
 
 uint32_t SurfaceKHR::getSurfaceFormatsCount() const
 {
-	return surfaceFormats.size();
+	return static_cast<uint32_t>(surfaceFormats.size());
 }
 
 VkResult SurfaceKHR::getSurfaceFormats(uint32_t *pSurfaceFormatCount, VkSurfaceFormatKHR *pSurfaceFormats) const
